feat: machine requests cache MAASENG-1834
