+++ .eslintrc.js
@@ -55,7 +55,7 @@ module.exports = {
           },
         ],
         "@typescript-eslint/consistent-type-imports": 2,
-        "@typescript-eslint/explicit-module-boundary-types": ["error"],
+        "@typescript-eslint/explicit-module-boundary-types": ["off"],
         "import/namespace": "off",
         "import/no-named-as-default": 0,
         "import/order": [
