+++ package.json
@@ -44,6 +44,8 @@
     "@canonical/react-components": "0.38.0",
     "@reduxjs/toolkit": "1.9.3",
     "@sentry/browser": "5.30.0",
+    "@tanstack/query-core": "4.29.11",
+    "@tanstack/react-query": "4.29.13",
     "classnames": "2.3.2",
     "clone-deep": "4.0.1",
     "date-fns": "2.29.3",
