--- src/app/machines/views/MachineList/MachineListTable/MachineListGroupCount/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./MachineListGroupCount";
