+++ src/testing/factories/state.ts
@@ -293,7 +293,6 @@ export const machineStateCount = define<MachineStateCount>({
   loaded: false,
   loading: false,
   stale: false,
-  params: null,
 });
 
 export const machineFilterGroup = define<FilterGroup>({
